requirements.txt
requests==2.25.1
python-dotenv==1.0.0
boto3==1.28.31
discord-webhook==0.13.0
